+++ packages/core/types/types/index.d.ts
@@ -1,0 +1,5 @@
+export * from './chain';
+export * from './common';
+export * from './manager';
+export * from './wallet';
+export * from './wallet-connect';