+++ packages/core/types/types/wallet-connect.d.ts
@@ -1,0 +1,7 @@
+import { Algo } from '@cosmjs/amino';
+export interface WCAccount {
+    algo: Algo;
+    address: string;
+    pubkey: string;
+    isNanoLedger?: boolean;
+}