+++ packages/cosmostation/types/wallet-connect/main-wallet.d.ts
@@ -1,0 +1,4 @@
+import { Wallet, WalletConnectWallet } from '@cosmos-kit/core';
+export declare class CosmostationMobileWallet extends WalletConnectWallet {
+    constructor(walletInfo: Wallet);
+}