+++ packages/react/types/modal/components/ConnectWalletButton.d.ts
@@ -1,0 +1,2 @@
+import { ConnectWalletType } from './types';
+export declare const ConnectWalletButton: ({ size, buttonText, isLoading, isDisabled, variant, leftIcon, rightIcon, onClickConnectBtn, }: ConnectWalletType) => JSX.Element;