+++ packages/react/types/modal/index.d.ts
@@ -1,0 +1,2 @@
+export * from './get-modal';
+export * from './get-modal-v2';