+++ packages/react/types/modal/theme.d.ts
@@ -1,0 +1,2 @@
+export declare const defaultTheme: Record<string, any>;
+export declare const noCssResetTheme: Record<string, any>;