+++ packages/omni/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.32.12](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/omni@0.32.11...@cosmos-kit/omni@0.32.12) (2022-12-15)
+
+**Note:** Version bump only for package @cosmos-kit/omni
+
+
+
+
+
+## 0.32.11 (2022-12-14)
+
+**Note:** Version bump only for package @cosmos-kit/omni
+
+
+
+
+
 ## [0.32.10](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/omni@0.32.9...@cosmos-kit/omni@0.32.10) (2022-12-11)
 
 **Note:** Version bump only for package @cosmos-kit/omni